+++ src/core/StelProjectorClasses.cpp
@@ -55,7 +55,7 @@ bool StelProjectorPerspective::forward(Vec3f &v) const
 
 bool StelProjectorPerspective::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	v[2] = std::sqrt(1.0/(1.0+v[0]*v[0]+v[1]*v[1]));
 	v[0] *= v[2];
 	v[1] *= v[2];
@@ -104,7 +104,7 @@ bool StelProjectorEqualArea::forward(Vec3f &v) const
 
 bool StelProjectorEqualArea::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	// FIXME: for high FoV, return false but don't cause crash with Mouse Pointer Coordinates.
 	const double dq = v[0]*v[0] + v[1]*v[1];
 	double l = 1.0 - 0.25*dq;
@@ -169,7 +169,7 @@ bool StelProjectorStereographic::forward(Vec3f &v) const
 
 bool StelProjectorStereographic::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const double lqq = 0.25*(v[0]*v[0] + v[1]*v[1]);
 	v[2] = lqq - 1.0;
 	v *= (1.0 / (lqq + 1.0));
@@ -230,7 +230,7 @@ bool StelProjectorFisheye::forward(Vec3f &v) const
 
 bool StelProjectorFisheye::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const double a = std::sqrt(v[0]*v[0]+v[1]*v[1]);
 	const double f = (a > 0.0) ? (std::sin(a) / a) : 1.0;
 	v[0] *= f;
@@ -272,16 +272,16 @@ bool StelProjectorHammer::forward(Vec3f &v) const
 	const float r = std::sqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2]);
 	const float alpha = std::atan2(v[0],-v[2]);
 	const float cosDelta = std::sqrt(1.f-v[1]*v[1]/(r*r));
-	float z = std::sqrt(1.+cosDelta*std::cos(alpha/2.f));
-	v[0] = 2.f*M_SQRT2*cosDelta*std::sin(alpha/2.f)/z * widthStretch;
-	v[1] = M_SQRT2*v[1]/r/z;
+	float z = std::sqrtf(1.f+cosDelta*std::cosf(alpha/2.f));
+	v[0] = 2.f*static_cast<float>(M_SQRT2)*cosDelta*std::sin(alpha/2.f)/z * widthStretch;
+	v[1] = static_cast<float>(M_SQRT2)*v[1]/r/z;
 	v[2] = r;
 	return true;
 }
 
 bool StelProjectorHammer::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const double zsq = 1.-0.25*0.25*v[0]*v[0]-0.5*0.5*v[1]*v[1];
 	const double z = zsq<0. ? 0. : std::sqrt(zsq);
 	const bool ret = 0.25*v[0]*v[0]+v[1]*v[1]<2.0; // This is stolen from glunatic
@@ -335,7 +335,7 @@ bool StelProjectorCylinder::forward(Vec3f &v) const
 
 bool StelProjectorCylinder::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const bool rval = v[1]<M_PI_2 && v[1]>-M_PI_2 && v[0]>-M_PI && v[0]<M_PI;
 	const double cd = std::cos(v[1]);
 	const double alpha=v[0];
@@ -384,7 +384,7 @@ bool StelProjectorMercator::forward(Vec3f &v) const
 
 bool StelProjectorMercator::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const bool rval = v[0]>-M_PI && v[0]<M_PI;
 	const double E = std::exp(v[1]);
 	const double h = E*E;
@@ -436,7 +436,7 @@ bool StelProjectorOrthographic::forward(Vec3f &v) const
 
 bool StelProjectorOrthographic::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const double dq = v[0]*v[0] + v[1]*v[1];
 	double h = 1.0 - dq;
 	if (h < 0) {
@@ -489,7 +489,7 @@ bool StelProjectorSinusoidal::forward(Vec3f &v) const
 
 bool StelProjectorSinusoidal::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
+	v[0] /= static_cast<double>(widthStretch);
 	const bool rval = v[1]<M_PI_2 && v[1]>-M_PI_2 && v[0]>-M_PI && v[0]<M_PI;
 	const double cd = std::cos(v[1]);
 	const double pcd = v[0]/cd;
@@ -531,8 +531,8 @@ bool StelProjectorMiller::forward(Vec3f &v) const
 
 bool StelProjectorMiller::backward(Vec3d &v) const
 {
-	v[0] /= widthStretch;
-	const double yMax=1.25f*asinh(tan(M_PI*2.0/5.0));
+	v[0] /= static_cast<double>(widthStretch);
+	const double yMax=1.25*asinh(tan(M_PI*2.0/5.0));
 	const bool rval = v[1]<yMax && v[1]>-yMax && v[0]>-M_PI && v[0]<M_PI;
 	const double lat = 1.25*atan(sinh(0.8*v[1]));
 	const double lng = v[0];
