+++ src/core/modules/NomenclatureMgr.cpp
@@ -414,16 +414,16 @@ void NomenclatureMgr::draw(StelCore* core)
 		// Early exit if the planet is not visible or too small to render the
 		// labels.
 		const Vec3d equPos = p->getJ2000EquatorialPos(core);
-		const double r = p->getEquatorialRadius() * p->getSphereScale();
+		const double r = p->getEquatorialRadius() * static_cast<double>(p->getSphereScale());
 		double angularSize = atan2(r, equPos.length());
-		double screenSize = angularSize * painter.getProjector()->getPixelPerRadAtCenter();
+		double screenSize = angularSize * static_cast<double>(painter.getProjector()->getPixelPerRadAtCenter());
 		if (screenSize < 50)
 			continue;
 		Vec3d n = equPos; n.normalize();
 		SphericalCap boundingCap(n, cos(angularSize));
 		if (!viewportRegion.intersects(boundingCap))
 			continue;
-		if (p->getVMagnitude(core) >= 20.)
+		if (p->getVMagnitude(core) >= 20.f)
 			continue;
 
 		// Render all the items of this planet.
@@ -456,7 +456,7 @@ void NomenclatureMgr::drawPointer(StelCore* core, StelPainter& painter)
 		painter.setColor(c[0],c[1],c[2]);
 		texPointer->bind();
 		painter.setBlending(true);
-		painter.drawSprite2dMode(screenpos[0], screenpos[1], 13.f, StelApp::getInstance().getTotalRunTime()*40.);
+		painter.drawSprite2dMode(static_cast<float>(screenpos[0]), static_cast<float>(screenpos[1]), 13.f, static_cast<float>(StelApp::getInstance().getTotalRunTime()*40.));
 	}
 }
