+++ src/scripting/StelScriptMgr.hpp
@@ -175,7 +175,7 @@ public slots:
 	//! because the script running at normal rate might set the simulation time rate
 	//! to be non-real time.
 	//! @param r rate, e.g. 2.0 = script runs at twice the normal rate
-	void setScriptRate(float r);
+	void setScriptRate(double r);
 	
 	//! Get the rate at which the script is running as a multiple of the normal
 	//! execution rate.
